chore(auth): Refactor authentication UX and add documentation
User authentication is now completely handled through pages/auth/[task.tsx].

BREAKING CHANGE: The login page path is now "app/auth/login" instead of
"app/auth/signin"
